ap_io2 (item submission)
Most of these updates were just for ap_io2 adding items to the database.
Feel free to try and break and give feedback.

Tooltip.css is a thing we can use on text as long as the tag has a
tooltip tag in it. Check out the example in ap_io2 for use case. You can
hover over Factor

createItem.js has the function that runs to validate the form before
sending it off to the database.

createItem.php has the actual database access and sending data.

I also like to take notes as I work, more for myself but sometimes they
help and have created a notes folder.