Update log in stuff
All pages can now redirect to the login page if the person isn't logged
in (using cookies). At the moment we have to add <?php include
'php/checkLogin.php';?> to every page for this to work properly, unless
we can find a more elegant way to add it (like maybe through a uniform
css file that we have to add anyways later).