More work on adding clients
Should now be able to add clients to the database and see them show up
on the main page in a table. It's a little wonky still. It will
eventually properly take you to the update page when you make a new
client. Right now it takes you there, and indicates there was a new
client. The back button here doesn't work like we would want, so that
might need some addressing. Tomorrow's goal is getting delete and update
functions working, as well as adding family members.