Add and Edit Client functional
Adding and editing a client are fully functional. Tomorrow it's on to
Adding and Updating Family Members. There's an issue when Updating a sql
value to Null, it causes a break in the query. So we may need to
disallow null, as updating without putting anything in a text box might
cause an issue. I will explore further tomorrow.

Added a few helper functions to the utilies .js and .php files as well.