Tasking txt files
Created a bunch of simple text files to try and keep track of what needs
to get done and what is done already. Each story folder has subtasks
inside it that have short descriptions of what they entail. If we need
more info or whatever we can update these as we see fit. I put the
client one in my in Progress folder and the Inventory one in Colton's
since we were working on those parts.