Most client ops working
Add, Edit and Remove Clients should be working
Add, Edit and Remove Family members should be working

Running into a weird issue with a SQL query returning no results if I
check the isDeleted flag. Might need to figure this out tomorrow.

Invoices are set to show in the client update window, but as no way to
create an invoice exists outside of command line, testing was limited.
(Did test with creating a single invoice, but nothing was added to it
and the view page doesn't exist yet, though it is linked to)