More work on appointments and order form creation started
Set up a screen for clients to pick a date after completing an order
form. Currently accessible after selecting a client name in the client
section. Started work on the order form, I'm pretty happy with where
it's going. Tomorrow I hope to have AJAX working on it so it updates
dynamically as values change. It might be a good idea to look at the
order form and create a txt file with a bunch of default values. Might
do that tomorrow or this weekend.