Client order forms now create invoice descriptions
Cleaned up a few things around there, but client order forms now
properly create invoice descriptions and followup to allow the client to
select a new appointment for next month. I've added a text file with
some test items (for the order form). Started working on the admin's
ability to view open orders. Going to need to ... guess what... add
another field to the invoice database for walkin status. I can't think
of a cleaner way to do it (other than making the status var negative and
taking abs whenever we need to use it, aside from checking if it's a
walkin).