Add Walkin-In Working
the awc page should be working now, can add an existing client or add a
new client. I ended up moving the javascript function
"getElementAndColorIt" to the utilities.js and removed it wherever I
found it in other scripts because we had it in script like 4 times (thus
me touching createDonation and createItem). I wasn't able to view my
invoices on the checkIn page, however. Not sure if there are -only-
walkins it doesn't show anything?