Part of review order form
new page: rof.php (review order form). Should pull up an order form for
review. It needs a bit more work and isn't fully functional, but the
hard part is in there and working. Had to make some database adjustments
to get it to work correctly, which kind of blows (had to add a new
primary key and a new flag to invoiceDescription), but it was the only
way I could think of getting it to work cleanly.