Add / Remove items from order
The admin can now remove or add items to an order at will. I'm not sure
if where I put it makes sense, as it muddles the screen that's just
supposed to be the order form. I think it would be better suited
somewhere else (maybe a copy of this page - ap_oo4)

Fixed the link from the client page to an invoice