Pushing changes for status variable
Status and order stuff all worked out on my end, I believe.
Client selection from login page now goes immediately to cp1.php.
cp1.php now has all active orders on it, ideally this is the page that
will be active on tablets, the registration table will select the
client's order from the table then hand them the tablet.

The status values are (in operation order): Available, Assigned, Active
(set for all orders for today when the checkin page loads), Arrived
(high and low), Ready to Review (high and low), Ready to Print (high and
low), Printed (High and low), completed. There is also an
'AdvanceInvoiceStatus()' function that currently adds 100 (using
defines) to a passed in value and returns the result. Use this function
to advance the status value whenever you need to advance it and
hopefully it should just... work. Status values already advance on my
end when a user submits their order (from arrived to ready to review)
and once reviewed it will now go to ready to print. So on your end,
Colton, just make sure the check in page sets all to active, sets
arrived appropriately, and then waits for printing. Printing should go
to ap_oo4.php (post vars are currently invoiceID, name, visitTime and
familySize, but if we need to limit that to fewer, we can and just do
the appropriate queries in oo4). I'm going to make ap_oo4 set the status
from ready to print to printed automatically, but that's not in this
checkin.