Invoices now set to print when print pressed
Instead of setting an invoice to printed when the page is loaded, the
print button now has to be pressed. Also updated some of the CSS
stylings for the view order page. It now displays everything a normal
page does, but hides some elements when the print preview is active.